Saner build directory
