+++ bin/lib/Logger.js
