+++ bin/lib/MediaWiki.js
