+++ bin/lib/Zim.js
