+++ bin/lib/mwoffliner.lib.js
