+++ bin/lib/redis.js
